+++ src/environment/index.ts
@@ -1,9 +1,4 @@
-export type {
-  UnpackData,
-  UnpackResult,
-  UnpackDFObject,
-  UnpackSuccess,
-} from './types.ts'
+export type { UnpackDFObject } from './types.ts'
 export {
   applyEnvironment,
   branch,
