+++ src/environment/tests/types.test.ts
@@ -1,36 +1,7 @@
-// deno-lint-ignore-file no-namespace ban-ts-comment
+// deno-lint-ignore-file no-namespace
 import { assertEquals, describe, it } from '../../test-prelude.ts'
-import { Result, Success } from '../../types.ts'
 import { withSchema } from '../../index.ts'
 import * as Subject from '../types.ts'
-
-namespace UnpackData {
-  const result = withSchema()(() => ({ name: 'foo' } as const))
-
-  type test = Expect<
-    Equal<Subject.UnpackData<typeof result>, { readonly name: 'foo' }>
-  >
-  type error = Expect<
-    // @ts-expect-error
-    Equal<Subject.UnpackData<typeof result>, { name: string }>
-  >
-}
-
-namespace UnpackResult {
-  const result = withSchema()(() => ({ name: 'foo' }))
-
-  type test = Expect<
-    Equal<Subject.UnpackResult<typeof result>, Result<{ name: string }>>
-  >
-}
-
-namespace UnpackSuccess {
-  const result = withSchema()(() => ({ name: 'foo' }))
-
-  type test = Expect<
-    Equal<Subject.UnpackSuccess<typeof result>, Success<{ name: string }>>
-  >
-}
 
 namespace UnpackAll {
   const dfA = withSchema()(() => ({ a: 1 } as const))
